Bug 558284 - replace com.spotify with org.mandas.docker-client
- make modifications so there is no API removal or changes requiring
  a major version bump
- add new IDockerContainerInfo2 interface that adds the new
  mounts() call
- add new IDockerContainerMountInfo to wrapper org.mandas
  CountainerMount
- make DockerContainerInfo implement IDockerContainerInfo2 as well
- add back the volumes() and volumesRW() methods to
  DockerContainerInfo and make them return null
- have DockerContainerInfo do the wrapping of ContainerMounts
  to IDockerContainerMounts
- add back getExecutionDriver() to DockerConnection
- change back DockerContainerExit.statusCode() back to return Integer
- change callers of DockerContainerInfo.mount() to use
  IDockerContainerMount and IDockerContainerInfo2

Reviewed-on: https://git.eclipse.org/r/163943
Tested-by: Linux Tools Bot <linuxtools-bot@eclipse.org>

+++ containers/org.eclipse.linuxtools.docker.core/src/org/eclipse/linuxtools/docker/core/IDockerContainerExit.java
@@ -14,6 +14,6 @@ package org.eclipse.linuxtools.docker.core;
 
 public interface IDockerContainerExit {
 
-	Long statusCode();
+	Integer statusCode();
 
 }
