+++ containers/org.eclipse.linuxtools.docker.core/src/org/eclipse/linuxtools/docker/core/IDockerContainerState.java
@@ -20,7 +20,7 @@ public interface IDockerContainerState {
 
 	Integer pid();
 
-	Long exitCode();
+	Integer exitCode();
 
 	Date startDate();
