+++ containers/org.eclipse.linuxtools.docker.core/src/org/eclipse/linuxtools/internal/docker/core/DockerContainerState.java
@@ -58,8 +58,8 @@ public class DockerContainerState implements IDockerContainerState {
 	}
 
 	@Override
-	public Long exitCode() {
-		return exitCode;
+	public Integer exitCode() {
+		return Integer.valueOf(exitCode.intValue());
 	}
 
 	@Override
